renaming `poseidon_hash.rs`& mv of `hash_to_field`
