+++ rln/src/poseidon_tree.rs
@@ -3,7 +3,7 @@
 // Implementation inspired by https://github.com/worldcoin/semaphore-rs/blob/d462a4372f1fd9c27610f2acfe4841fab1d396aa/src/poseidon_tree.rs (no differences)
 
 use crate::circuit::Fr;
-use crate::poseidon_hash::poseidon_hash;
+use crate::hashers::poseidon_hash;
 use cfg_if::cfg_if;
 use utils::merkle_tree::*;
